1.12.2-3.1.14
Changed over to the Forge Config Annotation System which causes the client to autogenerate a GUI for the config system.

Added a en_us.lang file to be used with the GUI.

The listen command now prints if it is off or on when toggled.

Built against FORGE build 2708

NOTE: you might have to copy over your config values.